Add script to find broken links inside the repo
Automate the process used to check for broken links inside the repo -
ain't nobody got time fo' dat.

After suggestions from @frankcash and @bryangarza, I've created a
multithreaded Python2 script to check for broken links, and report
them. When reporting, indicate the issues by file, to save time
tracking them down.

Average execution time during my tests has been just over 100 seconds
from start to finish, which compared to my previous synchronous script
is an incredible performance gain (~18x). This is done by creating
`2*NUM_CORES` Threads, which then check a URL each until complete. We
then report to the user the list of all files which contain links to
fix, as well as the links - making it less work to track down changes
needed.

We will easily be able to integrate this into Continuous Integration or
unit-testing-esq practice, as the script returns an error code of `1`
if we have any URLs that cannot be found.

Adapted a regex from SO (http://stackoverflow.com/a/25109573) instead
of the previous, homebaked, one as there was no point reinventing the
wheel. Adjusted the regex so it wouldn't match any of the badges, and
haven't seen any side effects related to it so far.

We make sure to spoof our UserAgent string so sites don't ignore our
requests by assuming we're just scraping the data.

Closes #55:
- [x] *Asynchronous in one manner*
- [x] *Run as unit tests*
  - Returns an error code of `1` if there are any broken URLs
- [x] *Ran through a code linter*
  - No PEP8 compliance issues found when run through `python2-flake8`
- [x] *Use a minimal amount of packages*
  - requires `python2-regex`, `python2-pycurl`, `python2-urllib` and
    `python2-termcolor`